requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
httpx==0.27.0
openai==1.30.1
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.31.0
